LayoutTests/imported/w3c/web-platform-tests/intersection-observer/v2/scaled-target-expected.txt



FAIL IntersectionObserver V2 test with scale applied to target. assert_equals: expected "[true]" but got "[null]"
